backend/requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
faiss-cpu
pillow
opencv-python
numpy
